+++ tests/up_db.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-SURREAL_CAPS_ALLOW_EXPERIMENTAL=graphql SURREAL_WEBSOCKET_RESPONSE_BUFFER_SIZE=0 surreal start memory --user root --pass root
+SURREAL_CAPS_ALLOW_EXPERIMENTAL=graphql,files SURREAL_WEBSOCKET_RESPONSE_BUFFER_SIZE=0 surreal start memory --user root --pass root